fix warp cheatcode

+++ fuzzing/testdata/contracts/cheat_codes/vm/warp.sol
@@ -1,10 +1,10 @@
 // This test ensures that the block timestamp can be set with cheat codes
 interface CheatCodes {
-    function warp(uint64) external;
+    function warp(uint256) external;
 }
 
 contract TestContract {
-    function test(uint64 x) public {
+    function test(uint256 x) public {
         // Obtain our cheat code contract reference.
         CheatCodes cheats = CheatCodes(0x7109709ECfa91a80626fF3989D68f67F5b1DD12D);
